🚚 Move: Sidepage 모아서 관리
