+++ src/components/SidePage/CardDetail/CardDetailStyled.ts
