+++ src/components/SidePage/CardDetail/CommentSection/CommentSection.tsx
@@ -4,10 +4,10 @@ import { Comment, CreateComments } from "src/utils/apiType";
 import useInputValue from "src/hooks/useInputValue";
 import { deleteComments, postComments } from "src/utils/api";
 import Button from "src/components/Button/Button";
-import { CurrentIdListType } from "../../../Card/Card";
-import EditingComment from "./EditingComment/EditingComment";
+import { CurrentIdListType } from "src/pages/Dashboard/components/Card/Card";
 import EmptyComment from "./EmptyComment";
 import * as S from "./CommentsSectionStyled";
+import EditingComment from "./EditComment/EditingComment";
 
 interface EditingState {
   isEditing: boolean;
